Make get_yelp useable as module or cli
Extracted functions for all of the user input and the file output,
called them all from a function that gets called when the script is run
from the command line. This leaves the main get_yelp function as
something that can be used from other pieces of code when passed the
required information. This is a step toward plugging it into a web
interface.